+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 3.41.1
+* Minor bug fix.
+
+## 3.41.0
+* Added SyncLicenseActivation() function which synchronizes the activation data with the Cryptlex servers.
+* Added SyncTrialActivation() function which synchronizes the trial activation data with the Cryptlex servers.
+
+## 3.40.0
+* Added baseValue property to the FeatureEntitlement object returned by GetFeatureEntitlement() function.
+
 ## 3.39.1
 * Minor bug fix.